transfer featured playlists to library
