+++ index.html
@@ -138,9 +138,6 @@
         <button>
           <i class="ri-filter-2-line"></i> Filter &lt;10:00
         </button>
-        <button>
-          <i class="ri-filter-2-line"></i> Filter Music
-        </button>
         <button class="checked" id="autoQueueBtn">
           <i class="ri-list-check-2"></i> Auto-Queue
         </button>
@@ -177,24 +174,12 @@
       </form>
       <ul id="suggestions"></ul>
       <div id="searchlist">
-        <h1 class="featuredPlaylists">
-          <svg xmlns="http://www.w3.org/2000/svg" width="3rem" height="3rem" viewBox="0 0 176 176"
-            style="transform:scale(0.7)">
-            <circle cx="88" cy="88" r="77" fill="var(--text)" />
-            <path fill="var(--bg)"
-              d="M88 46a42 42 0 1 1 0 84 42 42 0 0 1 0-84m0-4a46 46 0 1 0 0 92 46 46 0 0 0 0-92z" />
-            <path fill="var(--bg)" d="m72 111 39-24-39-22z" />
-          </svg>
-          <label>Featured Playlists</label>
-        </h1>
+        <h1> Your Search Results will show here</h1>
       </div>
     </section>
 
 
     <section id="library">
-
-      <h2 id="greeting">Welcome</h2>
-
       <div id="collections" align="center">
         <a class="collectionItem" id="discover" href="/list?collection=discover">
           <i class="ri-compass-3-line"></i>Discover
@@ -212,18 +197,33 @@
           <i class="ri-calendar-schedule-line"></i>Listen Later
         </a>
 
-        <a class="collectionItem" id="ftrdpls" href="/list?collection=listenLater">
-          <i class="ri-play-list-line"></i>Featured Playlists
+        <a class="collectionItem" id="ytm_pls" href="/list?collection=ytm_pls">
+          <i>
+            <svg xmlns="http://www.w3.org/2000/svg" width="3rem" height="3rem" viewBox="0 0 176 176"
+              style="transform:scale(0.7)">
+              <circle cx="88" cy="88" r="77" fill="var(--text)" />
+              <path fill="var(--bg)"
+                d="M88 46a42 42 0 1 1 0 84 42 42 0 0 1 0-84m0-4a46 46 0 1 0 0 92 46 46 0 0 0 0-92z" />
+              <path fill="var(--bg)" d="m72 111 39-24-39-22z" />
+            </svg>
+          </i>
+          Featured Playlists
         </a>
 
-        <a class="collectionItem" id="YPLs" href="/list?collection=listenLater">
-          <i class="ri-calendar-schedule-line"></i>Your Playlists
+        <a class="collectionItem" id="YPLs" href="/list?collection=ur_pls">
+          <i>
+            <svg xmlns="http://www.w3.org/2000/svg" version="1.0" viewBox="0 0 4172 4172" width="3.5rem"
+              height="3.5rem">
+              <path fill="var(--text)"
+                d="m1368 3037-55-10-23-6a369 369 0 0 1-57-19 552 552 0 0 1-266-246 437 437 0 0 1-31-74 590 590 0 0 1-18-245l5-25c7-35 21-77 35-105l9-19a522 522 0 0 1 679-236c1 5-1 77-3 91a1059 1059 0 0 1-24 119 274 274 0 0 1-19 53c-1 0-7-5-14-13-40-46-95-77-160-91a290 290 0 0 0-186 542 287 287 0 0 0 202 23c61-15 120-54 159-105a1108 1108 0 0 0 149-360 1296 1296 0 0 0 27-274 1164 1164 0 0 0-226-667 146 146 0 0 1-21-39l-4-9-4-11c-8-16-18-53-24-84-5-27-5-72 0-95 10-49 32-84 69-115 24-20 50-34 87-47a740 740 0 0 1 79-19c23-4 134-6 167-3a1364 1364 0 0 1 446 118l20 8 20 8 18 8a2232 2232 0 0 1 652 439 1008 1008 0 0 1 234 338c4 5 16 57 20 83 2 17 1 63-3 79-18 83-71 135-171 167-34 11-106 20-130 16-43-7-194-7-249 0-67 9-142 23-179 34l-34 10a974 974 0 0 0-94 33 1245 1245 0 0 0-170 84 1182 1182 0 0 0-405 414 529 529 0 0 1-347 244c-40 9-112 11-160 6zm1441-892 14-2c21-2 58-13 76-22 34-17 54-37 67-69 6-16 7-19 7-44 1-26 1-28-6-54-32-125-167-280-368-420-80-56-200-124-282-159l-26-12a1286 1286 0 0 0-124-47c-39-14-128-36-170-42l-18-3c-19-3-41-4-87-4-56 0-71 2-105 13-68 23-101 65-101 130a201 201 0 0 0 17 82c3 9 21 46 31 64 68 112 187 227 351 338a1827 1827 0 0 0 446 214 1084 1084 0 0 0 219 39c1 1 50-1 59-2z" />
+            </svg>
+          </i>Your Playlists
         </a>
-        <a class="collectionItem" id="ytpls" href="/list?collection=listenLater">
-          <i class="ri-calendar-schedule-line"></i>Youtube Playlists
+        <a class="collectionItem" id="ytpls" href="/list?collection=yt_pls">
+          <i class="ri-play-list-2-fill"></i>Saved Playlists
         </a>
-        <a class="collectionItem" id="channels" href="/list?collection=listenLater">
-          <i class="ri-calendar-schedule-line"></i>Channels
+        <a class="collectionItem" id="channels" href="/list?collection=channels">
+          <i class="ri-youtube-line"></i>Channels
         </a>
 
       </div>
@@ -278,7 +278,6 @@
       </div>
       <div>
         <h3>Search</h3>
-        <toggle-switch id="featuredPlaylistsSwitch" checked>Featured Playlists</toggle-switch>
         <toggle-switch id="defaultFilterSongs">Set YouTube Music as Default</toggle-switch>
         <toggle-switch id="suggestionsSwitch" checked>Display Suggestions</toggle-switch>
       </div>
