+++ src/stylesheets/style.css
@@ -191,6 +191,8 @@ a {
   --defaultText: #6dd;
 }
 
+
+
 .checked {
   --defaultBg: #505;
   --defaultText: #faf;
